requirements.txt
fastapi
uvicorn
pillow
python-dotenv
boto3
requests
torch
diffusers